download-artifact Compatibility between v1 and v2
When using download-artifact@v1, a directory denoted by the name of the artifact would be created if the path input was not provided. All of the contents would be downloaded to this directory.

With v2, there is no longer an extra directory that is created if the path input is not provided. All the contents are downloaded to the current working directory.